ensure that fake data has dtype float

+++ modopt/signal/wavelet.py
@@ -220,7 +220,7 @@ def get_mr_filters(data_shape, opt='', coarse=False,
     fake_data[tuple(zip(data_shape // 2))] = 1
 
     # Call mr_transform.
-    mr_filters = call_mr_transform(fake_data, opt=opt)
+    mr_filters = call_mr_transform(fake_data.astype(float), opt=opt)
 
     if trim:
         mr_filters = np.array([_trim_filter(f) for f in mr_filters])
